Implemeted get_best_wedding_items.
This sorts the list of wedding_items by ratio and then by name in descending order.
It then evaluates the best items to trade with the number of heart crystals the user
has input. These items are returned as a dictionary where the keys are the item
names and the values are how many the user should trade for to get the maximum profit.