Created 'Get All Needed Cards' test.
This test evaluates the worksheets that apply to the 'Needed Cards'
data. It tests that only the appropriate 5 worksheets (S1 Need, S2
Need, S3 Need, S4 Need, and WA Need) are evaluated.

The test also ensures that each worksheet has the same keys for
consistent comparisons to be made between all 5 of the worksheets.

Finally, the test tests whether or not there are 3 'needed' cards
that have been found.

NOTE: This test will require future changes as the 3 remaining cards
are found.