Created 'Get All Needed Cards' method.
This method returns all rows of 'needed' cards within a given
worksheet. This is a list of dictionaries where each dictionary
is a row. Each row has the keys 'Card Number', 'Card Character',
'Want', and 'Need'. Each of the returned rows  should have a value
within the 'Need' key, this is tested in 'test_get_all_needed_cards'.